feat: folders and main proj description
