+++ README.md
@@ -1,35 +1,12 @@
-# pymongo-api
+# Структура проекта
+* ./default-project - проект-исходник, который берется за основу для следующих папок.
 
-## Как запустить
+* ./mongo-sharding - настроено 2 шарда без репликасетов, 3 роутера, 3 конфига
 
-Запускаем mongodb и приложение
+* ./mongo-sharding-repl - дополнительно добавлены репликасеты в шарды
 
-```shell
-docker compose up -d
-```
+* ./sharding-repl-cache - финальный вариант проекта с кешированием запросов
 
-Заполняем mongodb данными
+* ./arch_scheme - схема архитектуры приложения
 
-```shell
-./scripts/mongo-init.sh
-```
-
-## Как проверить
-
-### Если вы запускаете проект на локальной машине
-
-Откройте в браузере http://localhost:8080
-
-### Если вы запускаете проект на предоставленной виртуальной машине
-
-Узнать белый ip виртуальной машины
-
-```shell
-curl --silent http://ifconfig.me
-```
-
-Откройте в браузере http://<ip виртуальной машины>:8080
-
-## Доступные эндпоинты
-
-Список доступных эндпоинтов, swagger http://<ip виртуальной машины>:8080/docs
+Внутри каждой папки есть README.md по запуску, а также скрипты. 1 папка = 1 независимое приложение.
